Move notifyOnDiscord function to commen-functions file
